util/mitmproxy_extends/requirements.txt
requests~=2.27.1
Flask~=2.0.3
Pillow~=9.2.0
PyYAML~=6.0
colorama~=0.4.4
mitmproxy~=8.0.0
ZA33_CMBSM
pycryptodome
watchdog
